+++ kurtosis-module/impl/participant_network/el/el_client_launcher.go
@@ -11,7 +11,8 @@ type ELClientLauncher interface {
 		enclaveCtx *enclaves.EnclaveContext,
 		serviceId services.ServiceID,
 		image string,
-		loglevel module_io.ParticipantLogLevel,
+		participantLogLevel string,
+		globalLogLevel module_io.GlobalClientLogLevel,
 		// If nil, then the node will be launched as a bootnode
 		bootnodeContext *ELClientContext,
 	) (